Crate google-deploymentmanager2_beta1

This documentation was generated from deploymentmanager crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

deployments (delete, get, insert and list)

manifests (get and list)

operations (get and list)

resources (get and list)

types (list)

Everything else about the deploymentmanager v2_beta1 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.operations().list(...).doit()
let r = hub.deployments().delete(...).doit()
let r = hub.operations().get(...).doit()
let r = hub.deployments().insert(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-deploymentmanager2_beta1 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-deploymentmanager2_beta1" as deploymentmanager2_beta1;
use deploymentmanager2_beta1::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use deploymentmanager2_beta1::Deploymentmanager;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Deploymentmanager::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.operations().list("project")
             .page_token("et")
             .max_results(-18)
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Uploads and Downlods

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Structs

Deployment
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert deployments (request)

get deployments (response)

list deployments (none)

delete deployments (none)

DeploymentDeleteMethodBuilder
Deletes a deployment and all of the resources in the deployment.

DeploymentGetMethodBuilder
Gets information about a specific deployment.

DeploymentInsertMethodBuilder
Creates a deployment and all of the resources described by the deployment manifest.

DeploymentListMethodBuilder
Lists all deployments for a given project.

DeploymentMethodsBuilder
A builder providing access to all methods supported on deployment resources. It is not used directly, but through the Deploymentmanager hub.

Deploymentmanager
Central instance to access all Deploymentmanager related resource activities

DeploymentsListResponse
A response containing a partial list of deployments and a page token used to build the next request if the request has been truncated.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list deployments (response)

Manifest
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list manifests (none)

get manifests (response)

ManifestGetMethodBuilder
Gets information about a specific manifest.

ManifestListMethodBuilder
Lists all manifests for a given deployment.

ManifestMethodsBuilder
A builder providing access to all methods supported on manifest resources. It is not used directly, but through the Deploymentmanager hub.

ManifestsListResponse
A response containing a partial list of manifests and a page token used to build the next request if the request has been truncated.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list manifests (response)

Operation
An operation resource, used to manage asynchronous API requests.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list operations (none)

delete deployments (response)

get operations (response)

insert deployments (response)

OperationError
[Output Only] If errors occurred during processing of this operation, this field will be populated.

This type is not used in any activity, and only used as part of another schema.

OperationErrorErrors
The array of errors encountered while processing this operation.

This type is not used in any activity, and only used as part of another schema.

OperationGetMethodBuilder
Gets information about a specific Operation.

OperationListMethodBuilder
Lists all Operations for a project.

OperationMethodsBuilder
A builder providing access to all methods supported on operation resources. It is not used directly, but through the Deploymentmanager hub.

OperationWarnings
[Output Only] If warning messages generated during processing of this operation, this field will be populated.

This type is not used in any activity, and only used as part of another schema.

OperationWarningsData
Metadata for this warning in 'key: value' format.

This type is not used in any activity, and only used as part of another schema.

OperationsListResponse
A response containing a partial list of operations and a page token used to build the next request if the request has been truncated.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list operations (response)

Resource
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get resources (response)

list resources (none)

ResourceGetMethodBuilder
Gets information about a single resource.

ResourceListMethodBuilder
Lists all resources in a given deployment.

ResourceMethodsBuilder
A builder providing access to all methods supported on resource resources. It is not used directly, but through the Deploymentmanager hub.

ResourcesListResponse
A response containing a partial list of resources and a page token used to build the next request if the request has been truncated.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list resources (response)

Type
A type supported by Deployment Manager.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list types (none)

TypeListMethodBuilder
Lists all Types for Deployment Manager.

TypeMethodsBuilder
A builder providing access to all methods supported on type resources. It is not used directly, but through the Deploymentmanager hub.

TypesListResponse
A response that returns all Types supported by Deployment Manager

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list types (response)

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.