Struct google-deploymentmanager2_beta1::Operation

pub struct Operation {
    pub status: String,
    pub insert_time: String,
    pub warnings: Vec<OperationWarnings>,
    pub http_error_message: String,
    pub target_id: String,
    pub target_link: String,
    pub start_time: String,
    pub creation_timestamp: String,
    pub id: String,
    pub name: String,
    pub self_link: String,
    pub operation_type: String,
    pub error: OperationError,
    pub progress: i32,
    pub end_time: String,
    pub http_error_status_code: i32,
    pub status_message: String,
    pub user: String,
}

An operation resource, used to manage asynchronous API requests.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list operations (none)

delete deployments (response)

get operations (response)

insert deployments (response)

Fields

status

[Output Only] Status of the operation. Can be one of the following: "PENDING", "RUNNING", or "DONE".

insert_time

[Output Only] The time that this operation was requested. This is in RFC 3339 format.

warnings

[Output Only] If warning messages generated during processing of this operation, this field will be populated.

http_error_message

[Output Only] If operation fails, the HTTP error message returned, e.g. NOT FOUND.

target_id

[Output Only] Unique target id which identifies a particular incarnation of the target.

target_link

[Output Only] URL of the resource the operation is mutating.

start_time

[Output Only] The time that this operation was started by the server. This is in RFC 3339 format.

creation_timestamp

[Output Only] Creation timestamp in RFC3339 text format.

id

[Output Only] Unique identifier for the resource; defined by the server.

name

[Output Only] Name of the operation.

self_link

[Output Only] Self link for the manifest.

operation_type

[Output Only] Type of the operation. Examples include "insert", or "delete"

error

[Output Only] If errors occurred during processing of this operation, this field will be populated.

progress

[Output Only] An optional progress indicator that ranges from 0 to 100. There is no requirement that this be linear or support any granularity of operations. This should not be used to guess at when the operation will be complete. This number should be monotonically increasing as the operation progresses.

end_time

[Output Only] The time that this operation was completed. This is in RFC3339 format.

http_error_status_code

[Output Only] If operation fails, the HTTP error status code returned, e.g. 404.

status_message

[Output Only] An optional textual description of the current status of the operation.

user

[Output Only] User who requested the operation, for example "user@example.com"